+++ src/code/nt/atkin.hpp
@@ -17,17 +17,17 @@ constexpr vecu atkin(u32 lim) {
 
   for (u32 i = 0; i < slim; ++i)
     for (u32 j = 0; j < slim; ++j) {
-      if (u64 n = 4_u64 * i * i + j * j; n > lim) break;
+      if (u64 n = 4_u64 * i * i + (u64)j * j; n > lim) break;
       else if (u32(n) % 12 == 1 || u32(n) % 12 == 5) vis[n].flip();
     }
   for (u32 i = 0; i < slim; ++i)
     for (u32 j = 0; j < slim; ++j) {
-      if (u64 n = 3_u64 * i * i + j * j; n > lim) break;
+      if (u64 n = 3_u64 * i * i + (u64)j * j; n > lim) break;
       else if (u32(n) % 12 == 7) vis[n].flip();
     }
   for (u32 i = 0; i < slim; ++i)
     for (u32 j = i - 1; ~j; --j) {
-      if (u64 n = 3_u64 * i * i - j * j; n > lim) break;
+      if (u64 n = 3_u64 * i * i - (u64)j * j; n > lim) break;
       else if (u32(n) % 12 == 11) vis[n].flip();
     }
   for (u32 i = 5; i <= slim; ++i)
